+++ packages/monaco-editor-wrapper/src/indexAllLanguages.ts
@@ -1,5 +1,15 @@
-export type { WebSocketConfigOptions, WorkerConfigOptions, LanguageExtensionConfig, WorkerCommunitcationConfig } from './index.js';
+export type {
+    WebSocketConfigOptions,
+    WorkerConfigOptions,
+    LanguageExtensionConfig,
+    WorkerCommunitcationConfig
+} from './index.js';
 
-export { CodeEditorConfig, MonacoEditorLanguageClientWrapper } from './index.js';
+export {
+    CodeEditorConfig,
+    MonacoEditorLanguageClientWrapper,
+    monaco,
+    vscode
+} from './index.js';
 
 import './indexLanguagesOnly.js';