+++ packages/monaco-editor-wrapper/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to npm module [monaco-editor-wrapper](https://www.npmjs.com/package/monaco-editor-wrapper) are documented in this file.
 
+## [1.4.0] - 2022-12-01
+
+- Export `vscode` (monaco-vscode-api) and `monaco` and remove getters
+- automaticLayout is configured as default
+- Fixed full configuration of editor and diff editor via `monacoEditorOptions` and `monacoDiffEditorOptions`
+- Changed the compile target and module to ES2022.
+
 ## [1.3.2] - 2022-11-25
 
 - Merged css and ttf helper functions. Now ttf is included in css removing unknown path errors.